make forward and backwards map
