+++ packages/vestjs-runtime/tsconfig.json
@@ -15,8 +15,8 @@
       "test-utils": ["./src/exports/test-utils.ts"],
       "ErrorStrings": ["./src/errors/ErrorStrings.ts"],
       "isolateInterfaces": ["./src/Isolate/isolateInterfaces.ts"],
+      "IsolateSerializer": ["./src/Isolate/IsolateSerializer.ts"],
       "IsolateSelectors": ["./src/Isolate/IsolateSelectors.ts"],
-      "IsolateParser": ["./src/Isolate/IsolateParser.ts"],
       "IsolateMutator": ["./src/Isolate/IsolateMutator.ts"],
       "IsolateKeys": ["./src/Isolate/IsolateKeys.ts"],
       "IsolateInspector": ["./src/Isolate/IsolateInspector.ts"],
